Update README.md
I have developed a Netflix replica web application featuring user authentication and a basic movie search functionality. The application allows users to sign up and sign in, with credentials securely managed using JavaScript and localStorage.

The structure of the website is built using HTML, while CSS were used to style and create a responsive and visually appealing interface. JavaScript was integrated to handle authentication logic and to implement a search feature for browsing a limited set of movies.